Move Generic/GenericDictionaryTests.cs
